Moved the toggler component into general
